# [Fix #12600] Support Prism as a Ruby parser