--- a/Rakefile
+++ b/Rakefile
@@ -24,7 +24,10 @@ Dir['tasks/**/*.rake'].each { |t| load t }
 desc 'Run RuboCop over itself'
 RuboCop::RakeTask.new(:internal_investigation)
 
-task default: %i[documentation_syntax_check spec ascii_spec internal_investigation]
+# The `ascii_spec` task has not been failing for a while, so it will not be run by default.
+# However, `ascii_spec` task will continue to be checked in CI. If there are any failures
+# originating from `ascii_spec` in CI, please run `bundle exec ascii_spec` to investigate.
+task default: %i[documentation_syntax_check spec prism_spec internal_investigation]
 
 require 'yard'
 YARD::Rake::YardocTask.new