--- a/lib/rubocop/config_validator.rb
+++ b/lib/rubocop/config_validator.rb
@@ -63,6 +63,10 @@ module RuboCop
       target_ruby.version
     end
 
+    def parser_engine
+      for_all_cops.fetch('ParserEngine', :parser_whitequark).to_sym
+    end
+
     def validate_section_presence(name)
       return unless @config.key?(name) && @config[name].nil?