--- a/lib/rubocop/rspec/support.rb
+++ b/lib/rubocop/rspec/support.rb
@@ -27,4 +27,5 @@ RSpec.configure do |config|
   config.include_context 'ruby 3.1', :ruby31
   config.include_context 'ruby 3.2', :ruby32
   config.include_context 'ruby 3.3', :ruby33
+  config.include_context 'ruby 3.4', :ruby34
 end