… .github/workflows/rubocop.yml
… .rubocop_todo.yml
… changelog/new_support_prism.md
… config/default.yml
… spec/rubocop/cop/layout/end_of_line_spec.rb
… spec/rubocop/cop/layout/heredoc_indentation_spec.rb
… spec/rubocop/cop/layout/indentation_style_spec.rb
… spec/rubocop/cop/layout/line_length_spec.rb
… spec/rubocop/cop/layout/space_around_keyword_spec.rb
… spec/rubocop/cop/layout/space_around_operators_spec.rb
… spec/rubocop/cop/layout/space_inside_hash_literal_braces_spec.rb
… spec/rubocop/cop/layout/trailing_whitespace_spec.rb
… spec/rubocop/cop/lint/ambiguous_operator_spec.rb
… spec/rubocop/cop/lint/ambiguous_regexp_literal_spec.rb
… spec/rubocop/cop/lint/circular_argument_reference_spec.rb
… spec/rubocop/cop/lint/deprecated_class_methods_spec.rb
… spec/rubocop/cop/lint/deprecated_constants_spec.rb
… spec/rubocop/cop/lint/erb_new_arguments_spec.rb
… spec/rubocop/cop/lint/non_deterministic_require_order_spec.rb
… spec/rubocop/cop/lint/numbered_parameter_assignment_spec.rb
… spec/rubocop/cop/lint/parentheses_as_grouped_expression_spec.rb
… spec/rubocop/cop/lint/redundant_dir_glob_sort_spec.rb
… spec/rubocop/cop/lint/redundant_require_statement_spec.rb
… spec/rubocop/cop/lint/refinement_import_methods_spec.rb
… spec/rubocop/cop/lint/require_parentheses_spec.rb
… spec/rubocop/cop/lint/syntax_spec.rb
… spec/rubocop/cop/lint/unified_integer_spec.rb
… spec/rubocop/cop/lint/useless_else_without_rescue_spec.rb
… spec/rubocop/cop/naming/block_forwarding_spec.rb
… spec/rubocop/cop/naming/heredoc_delimiter_case_spec.rb