Resolves #12600.

This is the initial basic implementation to support multiple parser engines. It is still in an experimental phase. There are still incompatibilities with Prism compared to the Parser gem, but as the RuboCop core, it is possible to begin providing support for Prism independently within the RuboCop core.

> [!IMPORTANT]
> To work this feature, the following patch for RuboCop AST needs to be released:
> https://github.com/rubocop/rubocop-ast/pull/277

## Setting the parser engine

RuboCop allows switching the backend parser by specifying either `parser_whitequark` or `parser_prism` for the `ParserEngine`.

Here are the parsers used as backends for each value:

- `ParserEngine: parser_whitequark` ... https://github.com/whitequark/parser
- `ParserEngine: parser_prism` ... https://github.com/ruby/prism (`Prism::Translation::Parser`)

By default, `parser_whitequark` is implicitly used.

`parser_whitequark` can analyze source code from Ruby 2.0 and above:

```yaml
AllCops:
  ParserEngine: parser_whitequark
```

`parser_prism` can analyze source code from Ruby 3.3 and above:

```yaml
AllCops:
  ParserEngine: parser_prism
  TargetRubyVersion: 3.3
```

`parser_prism` tends to perform analysis faster than `parser_whitequark`.

> [!CAUTION]
> Since the use of Prism is experimental, it is not included in RuboCop's runtime dependencies.
> If running through Bundler, please first add `gem 'prism'` to your Gemfile:
>
> ```ruby
> gem 'prism'
> ```

There are some incompatibilities with `parser_whitequark`, making it still considered experimental. For known issues, please refer to the Prism issues: https://github.com/ruby/prism/issues?q=is%3Aissue+is%3Aopen+label%3Arubocop

## Incompatibility

At the time of opening this PR, the following cops have incompatibility with Prism:

- `Gemspec/RequiredRubyVersion`
- `InternalAffairs/LocationLineEqualityComparison`
- `Layout/ClassStructure`
- `Layout/EmptyLines`
- `Layout/EndOfLine`
- `Layout/HeredocIndentation`
- `Layout/IndentationStyle`
- `Layout/LineLength`
- `Layout/SpaceAroundKeyword`
- `Layout/SpaceAroundOperators`
- `Layout/SpaceInsideHashLiteralBraces`
- `Layout/TrailingWhitespace`
- `Lint/AmbiguousOperator`
- `Lint/AmbiguousRegexpLiteral`
- `Lint/CircularArgumentReference`
- `Lint/DeprecatedClassMethods`
- `Lint/DeprecatedConstants`
- `Lint/ErbNewArguments`
- `Lint/NonDeterministicRequireOrder`
- `Lint/NumberedParameterAssignment`
- `Lint/ParenthesesAsGroupedExpression`
- `Lint/RedundantDirGlobSort`
- `Lint/RedundantRequireStatement`
- `Lint/RefinementImportMethods`
- `Lint/RequireParentheses`
- `Lint/Syntax`
- `Lint/UnifiedInteger`
- `Lint/UselessElseWithoutRescue`
- `Naming/BlockForwarding`
- `Naming/HeredocDelimiterCase`
- `Naming/HeredocDelimiterNaming`
- `Naming/VariableNumber`
- `Security/YamlLoad`
- `Style/ArgumentsForwarding`
- `Style/ArrayIntersect`
- `Style/BlockDelimiters`
- `Style/CollectionCompact`
- `Style/CommandLiteral`
- `Style/ComparableClamp`
- `Style/ConditionalAssignmentAssignInCondition`
- `Style/ConditionalAssignmentAssignToCondition`
- `Style/DataInheritance`
- `Style/DirEmpty`
- `Style/FileEmpty`
- `Style/FrozenStringLiteralComment`
- `Style/GuardClause`
- `Style/HashExcept`
- `Style/HashSyntax`
- `Style/HashTransformKeys`
- `Style/HashTransformValues`
- `Style/IfWithBooleanLiteralBranches`
- `Style/MultilineTernaryOperator`
- `Style/MultilineWhenThen`
- `Style/MutableConstant`
- `Style/NestedFileDirname`
- `Style/NumericLiterals`
- `Style/NumericPredicate`
- `Style/ObjectThen`
- `Style/QuotedSymbols`
- `Style/RedundantBegin`
- `Style/RedundantFreeze`
- `Style/RedundantHeredocDelimiterQuotes`
- `Style/RedundantParentheses`
- `Style/RescueModifier`
- `Style/SafeNavigation`
- `Style/SelectByRegexp`
- `Style/SingleLineMethods`
- `Style/SlicingWithRange`
- `Style/StringLiterals`
- `Style/TernaryParentheses`
- `Style/YamlFileRead`
- `Style/YodaCondition`

Some cop incompatibilities have been resolved in the Prism development line.

For known issues, please refer to the Prism issues: https://github.com/ruby/prism/issues?q=is%3Aissue+is%3Aopen+label%3Arubocop

### `Lint/Syntax` cop

The messages generated by Lint/Syntax depend on the parser engine used.

Parser gem:

```console
$ ruby -rparser/ruby33 -ve 'p Parser::Ruby33.parse("(")'
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) [x86_64-darwin22]
(string):1:2: error: unexpected token $end
```

Displays `unexpected token $end`.

Prism:

```console
$ ruby -rprism -rprism/translation/parser33 -ve 'p Prism::Translation::Parser33.parse("(")'
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) [x86_64-darwin22]
(string):1:2: error: expected a matching `)`
```

Displays `expected a matching )`.

There are differences in the messages between Parser gem and Prism, but since Prism can provide clearer messages in some cases, this incompatibility is accepted. In other words, the messages may vary depending on the parser engine.

## Test

To run tests with Prism, the command `bundle exec rake prism_spec` is provided. This task is also executed in GitHub Actions.

To run tests with Prism specifying files, set the environment variable `PARSER_ENGINE=parser_prism`:

```console
$ PARSER_ENGINE=parser_prism path/to/test_spec.rb
```

In the context of testing with Prism, two options for test cases are provided: `broken_on: :prism` and `unsupported_on: :prism`.
Both options are utilized to skip tests specifically for Prism.

### `broken_on: :prism`

Test cases failing due to Prism incompatibilities are marked with `broken_on: :prism`. This indicates an expectation for the issue to be resolved within Prism.

### `unsupported_on: :prism`

Prism is designed to parse Ruby versions 3.3+, which means that features unique to Ruby versions 2.0 through 3.2 are not supported.
Test cases falling into this category are marked with `unsupported_on: :prism`. This marker is used for cases that are testable with the Parser gem but not with Prism.

> [!NOTE]
> With `bundle exec rake`, `prism_spec` will be run instead of `ascii_spec`. The `ascii_spec` task has not been failing for a while, so it will not be run by default. However, `ascii_spec` task will continue to be checked in CI. If there are any failures originating from `ascii_spec` in CI, please run `bundle exec ascii_spec` to investigate.

-----------------

Before submitting the PR make sure the following are checked:

* [x] The PR relates to *only* one subject with a clear title and description in grammatically correct, complete sentences.
* [x] Wrote [good commit messages][1].
* [x] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [x] Added tests.
* [x] Ran `bundle exec rake default`. It executes all tests and runs RuboCop on its own code.
* [x] Added an entry (file) to the [changelog folder](https://github.com/rubocop/rubocop/blob/master/changelog/) named `{change_type}_{change_description}.md` if the new code introduces user-observable changes. See [changelog entry format](https://github.com/rubocop/rubocop/blob/master/CONTRIBUTING.md#changelog-entry-format) for details.

[1]: https://chris.beams.io/posts/git-commit/