@bbatsov 2024-02-29
@koic I've just cut `rubocop-ast` 1.31, so you can proceed with this PR. I also cut RuboCop 1.61, so we can have a separate release focused mostly on Prism support.