@koic 2024-03-01
@bbatsov Either option is fine. There might be points that could be improved upon in the future, but the basic implementation is complete.

One concern is whether the runtime dependency on Prism should be removed, similar to issue https://github.com/rubocop/rubocop-ast/issues/282. This might potentially become a blocker for upgrades, just as with RuboCop AST. 

The plan is to remove Prism from the rubocop.gemspec's runtime dependencies for the time being. Are there any concerns regarding this?

Currently, there is still the runtime dependency on Prism.