@koic 2024-03-01
For now, the runtime dependency on Prism has been removed from RuboCop, and the documentation has been updated to instruct on separate installation. This completes the current phase.