@bbatsov 2024-03-01
I think that's the best course of action for now. @kddnewton Please update at some point the Prism docs for using it with RuboCop as well to reflect the recent changes. I probably cut the next RuboCop release early next week.