@tisba 2024-03-01
This is super awesome, thank you for all your work! 🙏

I gave c123b96 a spin against our Rails monolith and it works great with only `Rails/UniqueValidationWithoutIndex` producing an error (`uninitialized constant Parser::Ruby33`). That's from the `rubocop-rails` gem and I'm aware that this cannot be ready yet. Do you keep track of that already somewhere, or do you want me to file an issue? Would like to help if I can!