@koic 2024-03-02
@tisba That needed to be resolved early on. I've opened https://github.com/rubocop/rubocop-rails/pull/1245. Thank you for the early feedback!