@eregon 2024-03-02
I was curious of the performance, so I tried this running rubocop on the whole ruby/spec, running on CRuby 3.3.0.
To install the gem since it's not released yet:
```
cd rubocop
bundle
bundle exec rake build
gem i pkg/rubocop-*.gem
```
With this diff in ruby/spec:
```diff
diff --git a/.rubocop.yml b/.rubocop.yml
index be32ce8900..5cf4bbf1b0 100644
--- a/.rubocop.yml
+++ b/.rubocop.yml
@@ -1,7 +1,8 @@
 inherit_from: .rubocop_todo.yml
 
 AllCops:
-  TargetRubyVersion: 3.0
+  TargetRubyVersion: 3.3
+  ParserEngine: parser_prism # also used commented out for the baseline
   DisplayCopNames: true
   Exclude:
     - command_line/fixtures/bad_syntax.rb
```
There are some expected offenses since I'm changing TargetRubyVersion to 3.3, e.g. `unnecessary require 'set'`.

baseline: TargetRubyVersion: 3.3, no ParserEngine set
```
$ time rubocop --cache=false .
4546 files inspected, 69 offenses detected, 68 offenses autocorrectable
rubocop --cache=false .  42.23s user 0.33s system 99% cpu 42.777 total

$ time rubocop --cache=false .
4546 files inspected, 69 offenses detected, 68 offenses autocorrectable
rubocop --cache=false .  42.23s user 0.34s system 99% cpu 42.794 total

$ time rubocop .
rubocop .  81.17s user 2.58s system 1077% cpu 7.774 total
$ time rubocop .
rubocop .  4.28s user 1.46s system 206% cpu 2.776 total
$ time rubocop .
rubocop .  4.25s user 1.51s system 209% cpu 2.753 total
```

Using the Prism translation layer: TargetRubyVersion: 3.3, ParserEngine: parser_prism:
```
$ time rubocop --cache=false .
4546 files inspected, 72 offenses detected, 68 offenses autocorrectable
rubocop --cache=false .  31.57s user 0.47s system 99% cpu 32.266 total

$ time rubocop --cache=false .
4546 files inspected, 72 offenses detected, 68 offenses autocorrectable
rubocop --cache=false .  31.50s user 0.40s system 99% cpu 32.069 total

$ time rubocop .
rubocop .  64.94s user 2.85s system 1011% cpu 6.702 total
$ time rubocop .
rubocop .  4.24s user 1.50s system 207% cpu 2.771 total
$ time rubocop .
rubocop .  4.20s user 1.49s system 204% cpu 2.776 total
```
So around a 25% speedup in that case for `--cache=false`.